refactor(desktop): centralize preload bridge contract
