+++ apps/desktop/src/main/__tests__/local-memory-ui-contract.test.ts
@@ -602,7 +602,7 @@ describe('local MEMORY.md Settings UI contract', () => {
   it('can restore the latest MEMORY.md backup through an explicit reversible action', async () => {
     const main = await readMainProcessCombinedSource();
     const preload = await readRepo('apps/desktop/src/preload/preload.ts');
-    const globalTypes = await readRepo('apps/desktop/src/global.d.ts');
+    const globalTypes = await readRepo('apps/desktop/src/preload/bridge-contract.d.ts');
     const service = await readRepo('apps/desktop/src/main/local-memory-service.ts');
     const src = await readSettingsCombinedSource();
     const pageBlock = src.match(/function MemorySettingsPage\([\s\S]*?function MemoryEntryList/)?.[0] ?? '';
@@ -700,7 +700,7 @@ describe('local MEMORY.md Settings UI contract', () => {
   it('opens the latest MEMORY.md backup only through a main-process validated path', async () => {
     const main = await readMainProcessCombinedSource();
     const preload = await readRepo('apps/desktop/src/preload/preload.ts');
-    const globalTypes = await readRepo('apps/desktop/src/global.d.ts');
+    const globalTypes = await readRepo('apps/desktop/src/preload/bridge-contract.d.ts');
     const service = await readRepo('apps/desktop/src/main/local-memory-service.ts');
     const src = await readSettingsCombinedSource();
     const pageBlock = src.match(/function MemorySettingsPage\([\s\S]*?function MemoryEntryList/)?.[0] ?? '';
@@ -725,7 +725,7 @@ describe('local MEMORY.md Settings UI contract', () => {
   it('opens a specific MEMORY.md backup candidate by kind without renderer-supplied paths', async () => {
     const main = await readMainProcessCombinedSource();
     const preload = await readRepo('apps/desktop/src/preload/preload.ts');
-    const globalTypes = await readRepo('apps/desktop/src/global.d.ts');
+    const globalTypes = await readRepo('apps/desktop/src/preload/bridge-contract.d.ts');
     const service = await readRepo('apps/desktop/src/main/local-memory-service.ts');
     const src = await readSettingsCombinedSource();
     const pageBlock = src.match(/function MemorySettingsPage\([\s\S]*?function MemoryEntryList/)?.[0] ?? '';
@@ -750,7 +750,7 @@ describe('local MEMORY.md Settings UI contract', () => {
   it('restores a specific MEMORY.md backup candidate by kind without renderer-supplied paths', async () => {
     const main = await readMainProcessCombinedSource();
     const preload = await readRepo('apps/desktop/src/preload/preload.ts');
-    const globalTypes = await readRepo('apps/desktop/src/global.d.ts');
+    const globalTypes = await readRepo('apps/desktop/src/preload/bridge-contract.d.ts');
     const service = await readRepo('apps/desktop/src/main/local-memory-service.ts');
     const src = await readSettingsCombinedSource();
     const pageBlock = src.match(/function MemorySettingsPage\([\s\S]*?function MemoryEntryList/)?.[0] ?? '';
