+++ apps/desktop/src/main/__tests__/session-sticky-model-contract.test.ts
@@ -79,7 +79,7 @@ describe('PR-SESSION-STICKY-MODEL-0 contract', () => {
   it('lets the user explicitly switch the current session model from the chat header', async () => {
     const main = await readMainProcessCombinedSource();
     const preload = await readFile(resolve(REPO_ROOT, 'apps/desktop/src/preload/preload.ts'), 'utf8');
-    const globalTypes = await readFile(resolve(REPO_ROOT, 'apps/desktop/src/global.d.ts'), 'utf8');
+    const globalTypes = await readFile(resolve(REPO_ROOT, 'apps/desktop/src/preload/bridge-contract.d.ts'), 'utf8');
     const renderer = await readRendererShellCombinedSource();
     const ui = await readModelSwitcherUiSource();
     const uiPrimitives = await readFile(resolve(REPO_ROOT, 'packages/ui/src/ui.tsx'), 'utf8');
