+++ apps/desktop/src/preload/attachment-ingest-payload.ts
